Changed deprecation notice to point to gulp 4.0
I couldn't figure out how the upgrade to orchestrator fixes this issue, but the upcoming gulp 4.0 release has exactly this functionality as part of replacing orchestrator with undertaker. I think this explanation is a bit more explicit by pointing right to the necessary new api methods, and may be more helpful to people trying to figure out the future best practice way to do this.